pyproject.toml
[tool.poetry]
name = "lxdapi"
version = "0.1.0"
description = "LXD API"
authors = ["Egor Ternovoy <cofob@riseup.net>"]

[tool.poetry.scripts]
lxdapi = "lxdapi.__main__:main"

[tool.poetry.dependencies]
python = ">=3.10,<4"

# FastAPI
fastapi = "^0.85" # Web-framework
pydantic = "^1.10" # Serialization-framework
python-multipart = "^0.0.5" # Multipart form files handling
uvicorn = { extras = ["standard"], version = "^0.18" } # ASGI web-server

# Database
SQLAlchemy = "^1.4" # Database-framework
asyncpg = "^0.26" # Async postgres driver
alembic = "^1.8" # Database migrations
python-dateutil = "^2.8" # Date handling
pytz = "^2022.6" # Timezone handling

# JWT
python-jose = "^3.3.0" # JWT handling

# CLI
click = "^8.1.3" # CLI framework

[tool.poetry.group.dev.dependencies]
black = "^22.8.0" # Code formatter
mypy = "^0.982" # Static type checking
isort = "^5.10.1" # Import sorting
flake8 = "^5.0.4" # Linter
flake8-docstrings = "^1.6.0" # Docstring linter
SQLAlchemy = {extras = ["mypy"], version = "^1.4.41"} # Mypy SA support
sqlalchemy2-stubs = "^0.0.2-alpha.27" # SA stubs
pre-commit = "^2.20.0" # Git pre commit hooks
pytest = "^7.1.3" # Testing framework
pytest-cov = "~4.0" # Coverage support for pytest
pytest-asyncio = "^0.19.0" # Async support for pytest
pytest-randomly = "~3.12" # Randomize test order
requests = "^2.28.1" # For FastAPI tests
faker = "^15.3.2" # For FastAPI tests

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
